requirements.txt
hmmlearn>=0.3.2
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.36
pandas-datareader>=0.10.0
alpaca-py>=0.20.0
pyyaml>=6.0
python-dotenv>=1.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pytest>=7.4.0
